pyproject.toml
[project]
name = "midtenn-lend-map"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
    "minio>=7.2.20",
    "pandas>=2.3.3",
    "prefect>=3.6.28",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "sqlmesh>=0.234.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
